bridge/requirements.txt
supabase
yt-dlp
python-dotenv
